Fix server starting errors in backend
