--- backend/src/socket/socket.types.ts
@@ -1,7 +1,0 @@
-import IUser from "../schema/types/user.type";
-
-declare module "socket.io" {
-  interface Socket {
-    user: IUser;
-  }
-}
